# Conflicts:
#	README.md
#	src/SB/Core/x/xCamera.h
#	src/SB/Core/x/xFX.cpp
#	src/SB/Core/x/xFX.h
#	src/SB/Core/x/xHud.cpp
#	src/SB/Core/x/xSnd.h
#	src/SB/Game/zHud.h
#	src/SB/Game/zNPCTypeBossPlankton.cpp
#	src/SB/Game/zTalkBox.cpp
#	src/SB/Game/zTaskBox.cpp

+++ src/SB/Game/zEntSimpleObj.cpp
@@ -237,7 +237,7 @@ void zEntSimpleObj_MgrUpdateRender(RpWorld* world, F32 dt)
                     }
                 }
                 model = ent->model;
-                model->Flags &= 0xFBFF;
+                model->Flags &= (u16)0xFBFF;
                 smgr->lastlod = picklod;
                 model->Bucket = smgr->lodBucket[picklod];
                 model->Data = (*model->Bucket)->OriginalData;
